quotes_site/pyproject.toml
[project]
name = "quotes-site"
version = "0.1.0"
description = ""
authors = [
    {name = "TarnavskyAndrew",email = "87767534+TarnavskyAndrew@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
mypy = "^1.17.1"
radon = "^6.0.1"
black = "^25.1.0"

